+++ src/main/scala/nl/biopet/utils/biowdl/samples/Rna3.scala
@@ -37,7 +37,7 @@ trait Rna3SingleEnd extends MultisamplePipeline {
 trait Rna3PairedEnd extends Rna3SingleEnd {
   override def samples: Map[String, Sample] =
     addReadgroup(super.samples,
-                 "wgs1",
+                 "rna3",
                  "lib1",
                  "rg1",
                  Map("R2" -> fixtureFile("samples", "rna3", "R2.fq.gz"),
